reuse cli e2e tests
