+++ e2e/cli/e2e-npm-commons/package.json
@@ -1,0 +1,11 @@
+{
+  "private": true,
+  "name": "@verdaccio/e2e-cli-npm-common",
+  "version": "1.0.2-next-8.0",
+  "dependencies": {
+    "@verdaccio/test-cli-commons": "workspace:2.0.0-next-8.0"
+  },
+  "scripts": {
+    "test": "echo no test"
+  }
+}
