+++ e2e/cli/e2e-npm9/package.json
@@ -4,6 +4,7 @@
   "version": "1.0.2-next-8.0",
   "dependencies": {
     "@verdaccio/test-cli-commons": "workspace:2.0.0-next-8.0",
+    "@verdaccio/e2e-cli-npm-common": "workspace:*",
     "npm": "9.9.3"
   },
   "scripts": {
